# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: MiroFish-Novel API v2 Baseline
  version: 0.1.0
  description: |
    Rebuild contract baseline for operation, error, and health resources.
servers:
  - url: http://127.0.0.1:5102
...
paths:
...
  /api/v2/worldlines/preparations:
    post:
      operationId: createPreparation
      tags: [worldline]
      summary: Create worldline preparation
      parameters:
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Preparation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/preparations/{preparationId}:
    get:
      operationId: getPreparation
      tags: [worldline]
      summary: Query worldline preparation
      parameters:
        - $ref: "#/components/parameters/PreparationId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Preparation detail
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/preparations/{preparationId}/agents:
    get:
      operationId: getPreparationAgents
      tags: [worldline]
      summary: Query preparation agents
      parameters:
        - $ref: "#/components/parameters/PreparationId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Preparation agents
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/preparations/{preparationId}/start:
    post:
      operationId: startPreparation
      tags: [worldline]
      summary: Start session from preparation
      parameters:
        - $ref: "#/components/parameters/PreparationId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "202":
          description: Preparation start accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions:
    get:
      operationId: listWorldlineSessions
      tags: [worldline]
      summary: List worldline sessions
      parameters:
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Session list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createWorldlineSession
      tags: [worldline]
      summary: Create worldline session
      parameters:
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Session accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}:
    get:
      operationId: getWorldlineSession
      tags: [worldline]
      summary: Query one worldline session
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/timeline:
    get:
      operationId: getWorldlineTimeline
      tags: [worldline]
      summary: Query worldline timeline
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/commands/advance:
    post:
      operationId: advanceWorldline
      tags: [worldline]
      summary: Advance worldline
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "202":
          description: Advance accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/commands/inject-variable:
    post:
      operationId: injectWorldlineVariable
      tags: [worldline]
      summary: Inject one worldline variable
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "202":
          description: Variable injection accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/commands/agent-actions:
    post:
      operationId: createAgentAction
      tags: [worldline]
      summary: Submit agent action
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "202":
          description: Agent action accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/commands/agent-dialogues:
    post:
      operationId: createAgentDialogue
      tags: [worldline]
      summary: Submit agent dialogue
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "202":
          description: Agent dialogue accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/agents/{agentId}:
    get:
      operationId: getSessionAgent
      tags: [worldline]
      summary: Query one session agent
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Session agent detail
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/agent-history:
    get:
      operationId: getAgentHistory
      tags: [worldline]
      summary: Query agent history
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Agent history
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/agent-actions:
    get:
      operationId: listAgentActions
      tags: [worldline]
      summary: Query agent actions
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Agent actions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/agent-memory:
    get:
      operationId: getAgentMemory
      tags: [worldline]
      summary: Query agent memory
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Agent memory
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/agent-memory-context:
    get:
      operationId: getAgentMemoryContext
      tags: [worldline]
      summary: Query agent memory context
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Agent memory context
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/agent-dialogues:
    get:
      operationId: listAgentDialogues
      tags: [worldline]
      summary: Query agent dialogues
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Agent dialogues
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/relations:
    get:
      operationId: getRelationHistory
      tags: [worldline]
      summary: Query relation history
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Relation history
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
...
  /api/v2/worldlines/sessions/{sessionId}/events:
    get:
      operationId: listWorldlineEvents
      tags: [worldline]
      summary: Query worldline events
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Worldline events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/auto-evolutions:
    post:
      operationId: startAutoEvolution
      tags: [worldline]
      summary: Start auto evolution
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Auto evolution accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/worldlines/sessions/{sessionId}/auto-evolutions/{autoEvolutionId}/events:
    post:
      operationId: streamAutoEvolution
      tags: [worldline]
      summary: Stream auto evolution events
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/AutoEvolutionId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Auto evolution event stream
          content:
            text/event-stream:
              schema:
                type: string
        "501":
          $ref: "#/components/responses/ErrorResponse"
components:
  parameters:
...
    PreparationId:
      name: preparationId
      in: path
      required: true
      schema:
        type: string
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
...
    AutoEvolutionId:
      name: autoEvolutionId
      in: path
      required: true
      schema:
        type: string
...
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Stable error payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ApiError:
      type: object
      required: [code, message, details, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        retryable:
          type: boolean
    ErrorResponse:
      type: object
      required: [trace_id, error]
      properties:
        trace_id:
          type: string
        error:
          $ref: "#/components/schemas/ApiError"
    OperationStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    OperationStepStatus:
      type: string
      enum: [pending, running, succeeded, failed, skipped]
    OperationStep:
      type: object
      required: [step_id, stage, status, progress_percent]
      properties:
        step_id:
          type: string
        stage:
          type: string
        status:
          $ref: "#/components/schemas/OperationStepStatus"
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        elapsed_ms:
          type: integer
          nullable: true
...
    OperationResource:
      type: object
      required:
        - operation_id
        - trace_id
        - status
        - progress_percent
        - retryable
        - steps
      properties:
        operation_id:
          type: string
        trace_id:
          type: string
        status:
          $ref: "#/components/schemas/OperationStatus"
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        retryable:
          type: boolean
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/OperationStep"